requirements.txt
flask==3.0.0
python-docx==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2
